Simple node moved from foreign parser
